Comments and cleanup
