+++ core/src/main/java/org/elasticsearch/search/fetch/innerhits/InnerHitsContext.java
