+++ core/src/main/java/org/elasticsearch/search/internal/ContextIndexSearcher.java
@@ -32,7 +32,6 @@ import org.elasticsearch.search.dfs.CachedDfSource;
 import org.elasticsearch.search.internal.SearchContext.Lifetime;
 import org.elasticsearch.common.lucene.search.ProfileQuery;
 import org.elasticsearch.search.profile.InternalProfileBreakdown;
-import org.elasticsearch.search.query.InternalProfiler;
 
 import java.io.IOException;
 import java.util.*;
@@ -78,6 +77,15 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
         }
     }
 
+    /**
+     * Profiling-aware wrapper of rewrite().  If profiling is enabled, the queryProfiler
+     * inside the searchContext is start/stop after the rewrite operation
+     *
+     * @param original Query to be rewritten
+     * @return         Rewritten query
+     *
+     * @throws IOException
+     */
     private Query doRewrite(Query original) throws IOException {
         if (profile) {
             searchContext.queryProfiler().startTime(original, InternalProfileBreakdown.TimingType.REWRITE);
@@ -106,6 +114,20 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
         }
     }
 
+    /**
+     * Profiling-aware wrapper of createNormalizedWeight().  If profiling is enabled,
+     * the queryProfiler inside the searchContext is start/stop after the normalization operation
+     *
+     * Note: if profiling is enabled, DFS is not used due to limitations of wrapping the
+     * IndexSearcher
+     *
+     * @param query         Query to build weight for
+     * @param needsScores   If this weight needs scores or not
+     * @param useDFS        If this query should use precomputed DFS tf/df
+     * @return              The normalized weight
+     *
+     * @throws IOException
+     */
     private Weight doCreateNormalizedWeight(Query query, boolean needsScores, boolean useDFS) throws IOException {
 
         // NOTE: DFS won't work!
@@ -119,6 +141,9 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
     @Override
     public Weight createWeight(Query query, boolean needsScores) throws IOException {
         if (profile) {
+            // createWeight() is called for each query in the tree, so we tell the queryProfiler
+            // each invocation so that it can build an internal representation of the query
+            // tree
             searchContext.queryProfiler().pushQuery(query);
 
             searchContext.queryProfiler().startTime(query, InternalProfileBreakdown.TimingType.WEIGHT);
@@ -151,10 +176,18 @@ public class ContextIndexSearcher extends IndexSearcher implements Releasable {
         }
     }
 
+    /**
+     * Return if this query is being profiled or not
+     */
     public boolean profile() {
         return profile;
     }
 
+    /**
+     * Sets if this query should be profiled or not
+     *
+     * @param profile True if this query should be profiled
+     */
     public void profile(boolean profile) {
         this.profile = profile;
     }
