+++ core/src/main/java/org/elasticsearch/search/internal/SearchContext.java
@@ -316,10 +316,23 @@ public abstract class SearchContext implements Releasable, HasContextAndHeaders 
 
     public abstract FetchSearchResult fetchResult();
 
+    /**
+     * Sets if this search context should be profiled or not
+     * @param profile True if the search should be profiled
+     */
     public abstract void profile(boolean profile);
 
+    /**
+     * Returns if this search context is being profiled
+     */
     public abstract boolean profile();
 
+    /**
+     * Returns the query profiler, which is used to store timings across the
+     * search context
+     *
+     * @return The query profiler
+     */
     public abstract InternalProfiler queryProfiler();
 
     /**
