+++ core/src/main/java/org/elasticsearch/search/profile/InternalProfileShardResults.java
@@ -1,5 +1,23 @@
-package org.elasticsearch.search.profile;
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
+package org.elasticsearch.search.profile;
 
 import com.google.common.collect.Maps;
 import org.elasticsearch.common.io.stream.StreamInput;
@@ -14,6 +32,10 @@ import java.util.*;
 
 import com.google.common.base.Function;
 
+/**
+ * This class is the internal representation of profiled results from all shards.  It is essentially
+ * a map of Shard -> Profile results, with some convenience methods and streamable/toxcontent
+ */
 public class InternalProfileShardResults implements ProfileResults, Streamable, ToXContent {
 
     private Map<SearchShardTarget, InternalProfileResult> results;
@@ -29,10 +51,22 @@ public class InternalProfileShardResults implements ProfileResults, Streamable, 
         }
     };
 
+    /**
+     * Add a shard's profile results to the map of all results
+     * @param shard             The shard where the results came from
+     * @param profileResults    The profile results for that shard
+     */
     public void addShardResult(SearchShardTarget shard, InternalProfileResult profileResults) {
         results.put(shard, profileResults);
     }
 
+    /**
+     * "Finalizes" the profile results by calculating the total time across all shards,
+     * then calling setGlobalTime() on each individual shard result.  This will recursively
+     * populate the relative times in all query nodes across all shards.
+     *
+     * This should be called after all shard results are added via addShardResult
+     */
     public void finalizeTimings() {
         long totalTime = 0;
         for (Map.Entry<SearchShardTarget, InternalProfileResult> entry : results.entrySet()) {
